Update package version and add chalk library for colored log outputs
Updated the version of the package from 1.5.0 to 1.7.0. The package files package.json and package-lock.json were updated accordingly. The 'chalk' library was also added as a new dependency in the package files. The library is used in src/index.js to stylize terminal string outputs, improving log message visibility by color-coding them according to type, such as error, warning, success, or info.